+++ src/components/common/Question.jsx
@@ -6,12 +6,15 @@ import { connect } from 'react-redux';
 import { isObject, isEqual, isEmpty } from 'lodash';
 import TextField from 'apollo-react/components/TextField';
 import { Checkmark } from '../svg';
+import { Edit } from '../svg';
 import Dropdown from './atoms/inputs/Dropdown';
 import TextArea from './atoms/inputs/TextArea';
-import UserLookup from './atoms/inputs/UserLookup';
 import { parseMomentDate } from '../../utils/DateUtils';
 import Multiselect from './atoms/inputs/Multiselect';
-import { setProposalAnswerData } from '../../redux/actions/proposal-actions';
+import {
+  setProposalAnswerData,
+  setEditQuestionData
+} from '../../redux/actions/proposal-actions';
 import { getUserData, getProposalDetails } from '../../redux/selectors';
 import MatomoHOC from '../HOC/MatomoHOC';
 import { getCountriesNameForCode, getCountryOptions } from '../../utils/utils';
@@ -20,12 +23,14 @@ import removeSpecialChars from '../../utils/pasteUtils';
 import Autocomplete from './atoms/inputs/AutoComplete';
 
 import DatePicker from 'apollo-react/components/DatePickerV2';
-import moment from 'moment'
+import moment from 'moment';
 import QuestionDatePicker from './atoms/inputs/QuestionDatePicker';
+import SFAnswerValidationWrapper from './SFAnswerValidationWrapper';
 // import DatePicker from './atoms/inputs/DatePicker';
 
 type State = {
-  selectedDay: string
+  selectedDay: string,
+  selectedRow: Boolean
 };
 
 type Props = {
@@ -44,7 +49,11 @@ type Props = {
   sfObject: string,
   sfField: string,
   milestone: any,
-  ismilestoneavailable: string
+  ismilestoneavailable: string,
+  setEditQuestionData: (data: Object) => void,
+  roleNames: Array<string>,
+  isCustomQuestion: boolean,
+  hasDifferentSFanswer: boolean
 };
 
 export class TaskRow extends Component<Props, State> {
@@ -52,7 +61,8 @@ export class TaskRow extends Component<Props, State> {
     super(props);
 
     this.state = {
-      selectedDay: ''
+      selectedDay: '',
+      selectedRow: false
     };
   }
 
@@ -83,16 +93,28 @@ export class TaskRow extends Component<Props, State> {
 
   handleTextChange = (textValue: string, lastAnswer: string) => {
     const { setProposalAnswer, proposalId, questionId, userData } = this.props;
-    let s1 = textValue.trim().split(' ').filter(v=>v.trim().length > 0);
-    let s2 = lastAnswer.trim().split(' ').filter(v=>v.trim().length > 0);
+    const s1 = textValue
+      .trim()
+      .split(' ')
+      .filter(v => v.trim().length > 0);
+    const s2 = lastAnswer
+      .trim()
+      .split(' ')
+      .filter(v => v.trim().length > 0);
     if (!isEmpty(textValue.replace(/\r?\n|\r| /g, ''))) {
       if (s1.length !== s2.length || s1.join(' ').trim() != s2.join(' ').trim())
-        setProposalAnswer(proposalId, questionId, String(textValue).trim(), userData);
+        setProposalAnswer(
+          proposalId,
+          questionId,
+          String(textValue).trim(),
+          userData
+        );
     } else if (!textValue.trim() && lastAnswer.trim()) {
       setProposalAnswer(proposalId, questionId, ' ', userData);
     }
 
     this.trackMatomoEventSubmitAnswer(textValue);
+    this.setSelectRow(false);
   };
 
   onClickChange = (selectedValue: string, lastAnswer: string) => {
@@ -102,6 +124,7 @@ export class TaskRow extends Component<Props, State> {
       setProposalAnswer(proposalId, questionId, selectedValue, userData);
 
     this.trackMatomoEventSubmitAnswer(selectedValue);
+    this.setSelectRow(false);
   };
 
   handleDayChange = (selectedDay: string, lastAnswer: Date) => {
@@ -109,7 +132,8 @@ export class TaskRow extends Component<Props, State> {
 
     this.setState({ selectedDay }, () => {
       if (
-        parseMomentDate(lastAnswer) !== parseMomentDate(selectedDay) &&
+        parseMomentDate(lastAnswer.trim()) !==
+          parseMomentDate(selectedDay.trim()) &&
         selectedDay
       )
         setProposalAnswer(proposalId, questionId, selectedDay, userData);
@@ -133,6 +157,14 @@ export class TaskRow extends Component<Props, State> {
     const { setQuestionToDisplayHistory, questionId } = this.props;
     setQuestionToDisplayHistory(questionId);
     this.trackMatomoEventAnswerHistory();
+  };
+
+  onChildInputFocus = event => {
+    this.setSelectRow(true);
+  };
+
+  setSelectRow = value => {
+    this.setState({ selectedRow: value });
   };
 
   trackMatomoEventSubmitAnswer = data => {
@@ -192,10 +224,15 @@ export class TaskRow extends Component<Props, State> {
   resetDate = () => {
     const { setProposalAnswer, proposalId, questionId, userData } = this.props;
     this.setState({ selectedDay: ' ' }, () => {
-      setProposalAnswer(proposalId, questionId, this.state.selectedDay, userData);
+      setProposalAnswer(
+        proposalId,
+        questionId,
+        this.state.selectedDay,
+        userData
+      );
       this.trackMatomoEventSubmitAnswer(' ');
-    })
-   }
+    });
+  };
 
   renderAnswer = (
     type: string,
@@ -219,9 +256,23 @@ export class TaskRow extends Component<Props, State> {
       else answerValue = answer.toString();
     }
 
-    if (sectionName === 'Proposal Team') 
-       return <Autocomplete sectionName={sectionName} onChange={this.handlePropsalChange} text={answerValue}/>
-      // return <UserLookup sectionName={sectionName} onChange={this.handleTextChange} text={answerValue} />;
+    if (sectionName === 'Proposal Team')
+      return (
+        <SFAnswerValidationWrapper
+          hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+          sfObject={sfObject}
+        >
+          <Autocomplete
+            sectionName={sectionName}
+            onFocus={e => this.setSelectRow(true)}
+            onBlur={e => this.setSelectRow(false)}
+            onChange={this.handlePropsalChange}
+            text={answerValue}
+          />
+        </SFAnswerValidationWrapper>
+      );
+
+    // return <UserLookup sectionName={sectionName} onChange={this.handleTextChange} text={answerValue} />;
 
     if (
       type === 'picklist' &&
@@ -235,81 +286,122 @@ export class TaskRow extends Component<Props, State> {
 
     switch (type) {
       case 'text':
-        answerValue = !(String(answerValue).trim()) ? '' : String(answerValue).trim();
+        answerValue = !String(answerValue).trim()
+          ? ''
+          : String(answerValue).trim();
         return (
-          <TextField
-            className="proposal-text-area"
-            placeholder="Click to answer"
-            onPaste={event => {
-              removeSpecialChars(event);
-              const txtareaheight =
-                event.target.scrollHeight > 300
-                  ? 300
-                  : event.target.scrollHeight;
-              event.target.style.height = `auto`;
-              event.target.style.height = `${txtareaheight + 2}px`;
-            }}
-            onChange={event => {
-              const txtareaheight =
-                event.target.scrollHeight > 300
-                  ? 300
-                  : event.target.scrollHeight;
-              event.target.style.height = `auto`;
-              event.target.style.height = `${txtareaheight + 2}px`;
-            }}
-            onBlur={e => this.handleTextChange(e.target.value, answerValue)}
-            defaultValue={answerValue}
-            sizeAdjustable
-            minHeight={40}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <TextField
+              className="proposal-text-area"
+              placeholder="Click to answer"
+              onPaste={event => {
+                removeSpecialChars(event);
+                const txtareaheight =
+                  event.target.scrollHeight > 300
+                    ? 300
+                    : event.target.scrollHeight;
+                event.target.style.height = `auto`;
+                event.target.style.height = `${txtareaheight + 2}px`;
+              }}
+              onChange={event => {
+                const txtareaheight =
+                  event.target.scrollHeight > 300
+                    ? 300
+                    : event.target.scrollHeight;
+                event.target.style.height = `auto`;
+                event.target.style.height = `${txtareaheight + 2}px`;
+              }}
+              onBlur={e => this.handleTextChange(e.target.value, answerValue)}
+              onFocus={e => this.onChildInputFocus(e)}
+              defaultValue={answerValue}
+              sizeAdjustable
+              minHeight={40}
+            />
+          </SFAnswerValidationWrapper>
         );
       case 'number':
-        answerValue = !(String(answerValue).trim()) ? '' : String(answerValue).trim();
+        answerValue = !String(answerValue).trim()
+          ? ''
+          : String(answerValue).trim();
         return (
-          <TextArea
-            className="proposal-text-area"
-            placeholder="Click to answer"
-            type="number"
-            onBlur={this.handleTextChange}
-            value={answerValue}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <TextArea
+              className="proposal-text-area"
+              placeholder="Click to answer"
+              type="number"
+              onBlur={this.handleTextChange}
+              onFocus={e => this.onChildInputFocus(e)}
+              value={answerValue}
+            />
+          </SFAnswerValidationWrapper>
         );
       case 'y/n':
         return (
-          <Dropdown
-            id="dd-proposal-answer"
-            placeholder="Click to answer"
-            items={optionsYN}
-            onClick={this.onClickChange}
-            value={answerValue}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <Dropdown
+              id="dd-proposal-answer"
+              placeholder="Click to answer"
+              items={optionsYN}
+              onClick={(val)=> this.onClickChange(val, answerValue)}
+              value={answerValue}
+              setSelectRow={this.setSelectRow}
+            />
+          </SFAnswerValidationWrapper>
         );
       case 'select':
         return (
-          <Dropdown
-            id="dd-proposal-answer"
-            placeholder="Click to answer"
-            items={finalOptions}
-            onClick={this.onClickChange}
-            value={answerValue}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <Dropdown
+              id="dd-proposal-answer"
+              placeholder="Click to answer"
+              items={finalOptions}
+              onClick={(val)=> this.onClickChange(val, answerValue)}
+              value={answerValue}
+              setSelectRow={this.setSelectRow}
+            />
+          </SFAnswerValidationWrapper>
         );
       case 'date':
         return (
-          <QuestionDatePicker 
-            value={answerValue} 
-            resetDate={this.resetDate}
-            handleDayChange={this.handleDayChange}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <QuestionDatePicker
+              value={answerValue}
+              resetDate={this.resetDate}
+              handleDayChange={this.handleDayChange}
+              onFocus={e => this.setSelectRow(true)}
+              onBlur={e => this.setSelectRow(false)}
+            />
+          </SFAnswerValidationWrapper>
         );
       case 'picklist':
         return (
-          <Multiselect
-            placeholder="Click to answer"
-            items={finalOptions}
-            onClick={this.onSelectValues}
-            value={answerValueComplex}
-          />
+          <SFAnswerValidationWrapper
+            hasDifferentSFanswer={this.props.hasDifferentSFanswer}
+            sfObject={sfObject}
+          >
+            <Multiselect
+              placeholder="Click to answer"
+              items={finalOptions}
+              onClick={this.onSelectValues}
+              value={answerValueComplex}
+              setSelectRow={this.setSelectRow}
+            />
+          </SFAnswerValidationWrapper>
         );
       default:
         return <div id="no-configuration">Click to answer</div>;
@@ -340,7 +432,12 @@ export class TaskRow extends Component<Props, State> {
       questionText,
       answerConfiguration,
       milestone,
-      ismilestoneavailable
+      ismilestoneavailable,
+      sectionName,
+      roleNames,
+      setEditQuestionData,
+      isCustomQuestion,
+      questionId: qId
     } = this.props;
     const questionId = answers.get('questionId');
     let lastAnswer;
@@ -350,10 +447,32 @@ export class TaskRow extends Component<Props, State> {
 
     if (lastAnswer) answerDate = parseMomentDate(lastAnswer.get('date'));
     return (
-      <div className="task-table-row">
+      <div
+        className={`task-table-row${
+          this.state.selectedRow ? ' selected-task-table-row' : ''
+        }`}
+      >
         <div className="question-text">
           {this.renderTags(milestone, ismilestoneavailable, lastAnswer)}
-          <p>{questionText}</p>
+          <p>
+            {questionText}
+            {isCustomQuestion && (
+              <span
+                onClick={() => {
+                  setEditQuestionData({
+                    questionText,
+                    section: sectionName,
+                    answerType: answerConfiguration.get('type'),
+                    roleNames,
+                    questionAnswered: lastAnswer ? true : false,
+                    questionId: qId
+                  });
+                }}
+              >
+                <Edit className="edit-icon" />
+              </span>
+            )}
+          </p>
         </div>
 
         <div>
@@ -382,5 +501,6 @@ const mapStateToProps = (state: Object) => ({
 });
 
 export default connect(mapStateToProps, {
-  setProposalAnswer: setProposalAnswerData
+  setProposalAnswer: setProposalAnswerData,
+  setEditQuestionData
 })(MatomoHOC(TaskRow));