+++ src/components/common/atoms/Buttons.jsx
@@ -6,20 +6,30 @@ type Props = {
   id?: string,
   children: any,
   onClick: Function,
-  className?: string
+  className?: string,
+  disabled?: any,
 };
 
 const defaultProps = {
   id: undefined,
-  className: undefined
+  className: undefined,
+  disabled: false
 };
 
-export const PrimaryButton = ({ id, onClick, children, className }: Props) => (
+export const PrimaryButton = ({
+  id,
+  onClick,
+  children,
+  className,
+  disabled
+}: Props) => (
   <button
     id={id}
+    disabled={disabled || false}
     type="button"
     className={classnames('primary-button', className)}
     onClick={onClick}
+    disabled={disabled}
   >
     {children}
   </button>